Fix bug in getTemplateData
